Add files via upload
A random background image can be used for various purposes on a webpage in different ways. Sometimes it is used as a slideshow and sometimes one image is displayed at a time and a new image appears when the page is reloaded. You can see this feature on Bing’s homepage. In this tutorial, we will create a random background image functionality on page refresh using jQuery and CSS.